# api-rest-amazonia/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/auth/login": {
      "post": {
        "description": "🔓 **Acceso público — sin token requerido.**\n\nValida email y contraseña y devuelve un token JWT de acceso.\n\n**Uso del token:** incluye el `accessToken` recibido en todas las peticiones protegidas:\n`Authorization: Bearer <accessToken>`\n\n**Seguridad:**\n- El tiempo de respuesta es constante independientemente de si el email existe, para prevenir enumeración de usuarios (OWASP A07).\n- **Rate limiting:** máximo 5 intentos por IP cada 60 segundos. Superar este límite devuelve `429`.",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Login exitoso. Devuelve el token JWT y su tiempo de expiración.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Respuesta en caso de parametros invalidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationExceptionDto"
                }
              }
            }
          },
          "401": {
            "description": "Credenciales inválidas, cuenta desactivada o acceso expirado (investigadores)."
          },
          "429": {
            "description": "Demasiadas solicitudes desde esta IP. Espera antes de reintentar.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponseDto"
                }
              }
            }
          }
        },
        "summary": "Iniciar sesión",
        "tags": [
          "Auth",
          "Auth — Público"
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "description": "🔒 **Requiere token válido (cualquier rol).**\n\nConfirma el cierre de sesión del usuario autenticado.\n\n**Importante:** los tokens JWT son stateless — la invalidación real ocurre eliminando el token en el cliente (localStorage, memoria, cookie). Este endpoint sirve como confirmación semántica del logout.\n\nEn una futura versión se puede extender con una blacklist de tokens para invalidación inmediata en servidor.",
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Sesión cerrada. El cliente debe eliminar el token localmente."
          },
          "401": {
            "description": "Token inválido, expirado o no proporcionado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Cerrar sesión",
        "tags": [
          "Auth",
          "Auth — Usuario Autenticado"
        ]
      }
    },
    "/api/auth/me": {
      "get": {
        "description": "🔒 **Requiere token válido (cualquier rol).**\n\nDevuelve los datos del perfil del usuario autenticado a partir del token JWT.\n\n**Campos devueltos:** `id`, `email`, `nombre`, `rol` (1=Superadmin, 2=Admin, 3=Investigador), `activo`, `fechaExpiracion` (null para Superadmin y Admin), `createdAt`, `updatedAt`.\n\n**Nunca se devuelve:** `passwordHash` ni ninguna credencial.",
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Perfil completo del usuario autenticado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Token inválido, expirado o no proporcionado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Obtener perfil propio",
        "tags": [
          "Auth",
          "Auth — Usuario Autenticado"
        ]
      },
      "put": {
        "description": "🔒 **Requiere token válido (cualquier rol).**\n\nPermite al usuario actualizar su propio perfil.\n\n**Restricción:** solo se puede modificar el `nombre`. El `rol` y el estado `activo` no son editables desde este endpoint — esos campos son gestionados exclusivamente por Admin.",
        "operationId": "AuthController_updateMe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePerfilPropioDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Perfil actualizado exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Respuesta en caso de parametros invalidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationExceptionDto"
                }
              }
            }
          },
          "401": {
            "description": "Token inválido, expirado o no proporcionado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Actualizar perfil propio",
        "tags": [
          "Auth",
          "Auth — Usuario Autenticado"
        ]
      }
    },
    "/api/auth/change-password": {
      "post": {
        "description": "🔒 **Requiere token válido (cualquier rol).**\n\nPermite cambiar la contraseña verificando primero la contraseña actual.\n\n**Reglas de la nueva contraseña:**\n- Mínimo 8 caracteres\n- Al menos una mayúscula\n- Al menos un número\n- Al menos un símbolo (por ejemplo: ! @ # $)\n\n**Nota:** las sesiones activas con el token anterior siguen siendo válidas hasta que expiren de forma natural (TTL del JWT). Para invalidación inmediata se requiere implementar blacklist de tokens (backlog).",
        "operationId": "AuthController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contraseña cambiada exitosamente."
          },
          "400": {
            "description": "Respuesta en caso de parametros invalidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationExceptionDto"
                }
              }
            }
          },
          "401": {
            "description": "401: token inválido/expirado, o la contraseña actual es incorrecta."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Cambiar contraseña propia",
        "tags": [
          "Auth",
          "Auth — Usuario Autenticado"
        ]
      }
    },
    "/api/auth/register": {
      "post": {
        "description": "🔒 **Requiere rol: Admin o Superadmin.**\n\nCrea un nuevo usuario con rol Admin (2) o Superadmin (1).\n\n**Reglas de creación de usuarios:**\n- Un `Admin` solo puede crear usuarios con rol `Admin` (2). Intentar crear un `Superadmin` devuelve `403`.\n- Un `Superadmin` puede crear usuarios con cualquier rol.\n- Los `Investigadores` se crean únicamente a través del flujo de solicitud de acceso (`POST /auth/solicitar-acceso` → aprobación).\n\n**Reglas de contraseña:**\n- Mínimo 8 caracteres, al menos una mayúscula, un número y un símbolo.\n\n**Rate limiting:** máximo 10 creaciones por IP cada 60 segundos.",
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterUsuarioDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario creado exitosamente. Devuelve el perfil sin contraseña.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Respuesta en caso de parametros invalidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationExceptionDto"
                }
              }
            }
          },
          "401": {
            "description": "Token inválido, expirado o no proporcionado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "403: rol insuficiente, o Admin intentando crear Superadmin."
          },
          "409": {
            "description": "Respuesta en caso de un conflicto entre datos repetidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Demasiadas solicitudes desde esta IP. Espera antes de reintentar.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Crear nuevo usuario (Admin o Superadmin)",
        "tags": [
          "Auth",
          "Auth — Admin"
        ]
      }
    },
    "/api/auth/usuarios": {
      "get": {
        "description": "🔒 **Requiere rol: Admin o Superadmin.**\n\nDevuelve el listado paginado de usuarios con filtros opcionales.\n\n**Filtros disponibles:**\n- `?rol=1` — Superadmin | `?rol=2` — Admin | `?rol=3` — Investigador\n- `?activo=true` — solo activos | `?activo=false` — solo inactivos\n- `?search=texto` — busca por nombre o email (parcial, insensible a mayúsculas)\n\n**Ejemplos:** `?rol=3&activo=true` → investigadores activos · `?search=Juan` → usuarios que contienen \"Juan\" en nombre o email\n\n**Paginación:** `?page=1&limit=10`. Límite máximo: 100.",
        "operationId": "AuthController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Número de la página (empieza en 1). Si no se envía, se toma por defecto el valor 1.",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Máximo 100",
            "schema": {
              "maximum": 100,
              "default": 10,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "rol",
            "required": false,
            "in": "query",
            "description": "1=Superadmin, 2=Admin, 3=Investigador",
            "schema": {
              "enum": [
                1,
                2,
                3
              ],
              "type": "number"
            }
          },
          {
            "name": "activo",
            "required": false,
            "in": "query",
            "description": "true=activos, false=inactivos",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Busca en nombre y email",
            "schema": {
              "maxLength": 100,
              "example": "María",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Listado paginado de usuarios.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginationResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Token inválido, expirado o no proporcionado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "El usuario no tiene el rol necesario para este recurso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Listar usuarios con filtros y paginación",
        "tags": [
          "Auth",
          "Auth — Admin"
        ]
      }
    },
...
    "/api/auth/solicitar-acceso": {
      "post": {
        "description": "🔓 **Acceso público — sin token requerido.**\n\nEnvía una solicitud de acceso temporal como investigador a la plataforma.\n\n**Flujo completo:**\n1. El solicitante envía este formulario con su nombre, email, institución y propósito de investigación.\n2. Un Admin revisa la solicitud en `GET /auth/solicitudes`.\n3. Si se aprueba (`PATCH /auth/solicitudes/:id/aprobar`), se crea automáticamente una cuenta de Investigador con acceso temporal hasta la fecha definida por el Admin.\n4. El investigador recibe las credenciales y puede hacer login con `POST /auth/login`.\n\n**Restricción:** no se puede enviar una nueva solicitud mientras haya una pendiente con el mismo email. Devuelve `409`.\n\n**Rate limiting:** máximo 3 solicitudes por IP cada 60 segundos.",
        "operationId": "SolicitudesController_solicitar",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearSolicitudDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Solicitud enviada exitosamente. Será revisada por un administrador."
          },
          "400": {
            "description": "Respuesta en caso de parametros invalidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationExceptionDto"
                }
              }
            }
          },
          "409": {
            "description": "409: ya existe una solicitud pendiente con ese correo electrónico."
          },
          "429": {
            "description": "Demasiadas solicitudes desde esta IP. Espera antes de reintentar.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponseDto"
                }
              }
            }
          }
        },
        "summary": "Solicitar acceso como investigador",
        "tags": [
          "Solicitudes",
          "Auth — Público"
        ]
      }
    },
...
  },
  "info": {
    "title": "Kaa Iya — Backend API",
    "description": "API REST de la plataforma **Kaa Iya** (\"Espíritu del Bosque\"), para gestionar y visibilizar iniciativas sostenibles en la Amazonía boliviana.\n\n**Base URL:** `/api`\n\n**Autenticación:** Bearer JWT. Obtén tu token en `POST /auth/login` e inclúyelo en todas las peticiones protegidas: `Authorization: Bearer <token>`",
    "version": "1.0.0",
    "contact": {
      "name": "Equipo Kaa Iya — UCB San Pablo",
      "url": "https://www.ucb.edu.bo",
      "email": "catedra.amazonia@ucb.edu.bo"
    },
    "license": {
      "name": "UNLICENSED",
      "url": ""
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3333",
      "description": "Local development"
    },
    {
      "url": "https://kaaiya-backend.onrender.com",
      "description": "Producción (Render/Railway — ajustar al dominio real)"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Ingresa el token JWT obtenido de POST /auth/login"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Correo electrónico del usuario",
            "example": "admin@kaaiya.bo"
          },
          "password": {
            "type": "string",
            "description": "Contraseña del usuario",
            "example": "MiPassword1!"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "TokenResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "Token JWT de acceso",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "tipo": {
            "type": "string",
            "description": "Tipo de token",
            "example": "Bearer"
          },
          "expiresIn": {
            "type": "string",
            "description": "Tiempo de expiración del token",
            "example": "24h"
          }
        },
        "required": [
          "accessToken",
          "tipo",
          "expiresIn"
        ]
      },
...
      "CommonResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Mensaje de respuesta"
          }
        },
        "required": [
          "message"
        ]
      },
...
    }
  }
}